requirements.txt
streamlit
fastapi
uvicorn
python-multipart
pyyaml
numpy
requests
pypdf
sentence-transformers
faiss-cpu
langchain-text-splitters
tiktoken
rank_bm25
pdfplumber
easyocr
pydantic==2.9.2
gpustat
reportlab
plotly
pillow
pytesseract
tqdm
watchdog
opencv-python-headless
pandas
pytest
httpx
